+++ CHANGELOG.md
@@ -11,6 +11,10 @@
 
 - Kaiju Upgrade Tech replaces the side guns due to the original tech now being in the base game
 
+### Fixed
+
+- Torpedo upgrades not working correctly due to changes in PA 120679
+
 ## v5.63.0 - 2024-03-21
 
 ### Added